+++ internal/service/blob/lfs_pointers.go
@@ -1,0 +1,57 @@
+package blob
+
+import (
+	"fmt"
+
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+)
+
+func (s *server) GetLFSPointers(req *pb.GetLFSPointersRequest, stream pb.BlobService_GetLFSPointersServer) error {
+	ctx := stream.Context()
+
+	if err := validateGetLFSPointersRequest(req); err != nil {
+		return grpc.Errorf(codes.InvalidArgument, "GetLFSPointers: %v", err)
+	}
+
+	client, err := s.BlobServiceClient(ctx)
+	if err != nil {
+		return err
+	}
+
+	clientCtx, err := rubyserver.SetHeaders(ctx, req.GetRepository())
+	if err != nil {
+		return err
+	}
+
+	rubyStream, err := client.GetLFSPointers(clientCtx, req)
+	if err != nil {
+		return err
+	}
+
+	return rubyserver.Proxy(func() error {
+		resp, err := rubyStream.Recv()
+		if err != nil {
+			md := rubyStream.Trailer()
+			stream.SetTrailer(md)
+			return err
+		}
+		return stream.Send(resp)
+	})
+}
+
+func validateGetLFSPointersRequest(req *pb.GetLFSPointersRequest) error {
+	if req.GetRepository() == nil {
+		return fmt.Errorf("empty Repository")
+	}
+
+	if len(req.GetBlobIds()) == 0 {
+		return fmt.Errorf("empty BlobIds")
+	}
+
+	return nil
+}