+++ internal/service/blob/server.go
@@ -1,10 +1,16 @@
 package blob
 
-import pb "gitlab.com/gitlab-org/gitaly-proto/go"
+import (
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
 
-type server struct{}
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+)
+
+type server struct {
+	*rubyserver.Server
+}
 
 // NewServer creates a new instance of a grpc BlobServer
-func NewServer() pb.BlobServiceServer {
-	return &server{}
+func NewServer(rs *rubyserver.Server) pb.BlobServiceServer {
+	return &server{rs}
 }